Moved User Agent parser to a new BrowserMobProxyUtil class
